connexion-example/requirements.txt
connexion[flask, swagger-ui]
bcrypt
Flask-Sqlalchemy
# bcrypt==3.2.2
# pyjwt==2.0.1
# Flask==2.2.3 

# connexion[swagger-ui]==2.13.1
# Flask-Sqlalchemy==3.0.3
# bcrypt==3.2.2
# pyjwt==2.0.1
# Flask==2.2.3 

# python3.11 -m venv .venv
# source .venv/bin/activate